\name{msplot}
\alias{msplot}
\title{Drawing the Results of Curve Fitting for Pvclust Object}
\description{draws the results of curve fitting for pvclust object.
}
\usage{
msplot(x, edges=NULL, ...)
}
\arguments{
  \item{x}{object of class \code{pvclust}.}
  \item{edges}{numeric vector of edge numbers to be plotted.}
  \item{...}{other parameters to be used in the function.}
}
\seealso{\code{\link{plot.msfit}}}
\author{Ryota Suzuki \email{suzuki@ef-prime.com}}
\keyword{hplot}
